# openapi.yaml
openapi: 3.0.3
info:
  title: API Newtral
  description: API implementada para el proyecto de Newtral.
  contact:
    email: marcialfco@live.com
  version: 0.0.1
servers:
  - url: http://localhost:3000
...
paths:
...
  /politicians/{politicianId}:
    get:
      tags:
        - politicians
      summary: Obtener político por id
      description: Devuelve un único político
      operationId: getPoliticianById
      parameters:
        - name: politicianId
          in: path
          description: ID del político a obtener
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliticianResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      tags:
        - politicians
      summary: Actualiza los datos de un político
      description: ''
      operationId: updatePetWithForm
      parameters:
        - name: politicianId
          in: path
          description: ID del político a modificar
          required: true
          schema:
            type: string
      requestBody:
        description: Atributos del político que se van a modificar
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Politician'
        required: true
      responses:
        '200':
          description: Operación correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliticianPatchResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - politicians
      summary: Elimina un político
      description: Elimina un político
      operationId: deletePolitician
      parameters:
        - name: politicianId
          in: path
          description: ID del político a eliminar
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliticianPatchResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    Politician:
      type: object
      properties:
          nombre: 
            type: string
            example: "Marcial Francisco Parrilla Socas"
          partido: 
            type: string
            example: "Partido Socialista Obrero Español"
          partido_para_filtro:
            type: string
            example: "PSOE"
          genero: 
            type: string
            example: "Hombre"
          cargo_para_filtro: 
            type: string
            example: "Alcalde"
          cargo: 
            type: string
            example: "Alcalde"
          institucion: 
            type: string
            example: "Ayuntamiento de Teguise"
          ccaa: 
            type: string
            example: "Canarias"
          sueldobase_sueldo: 
            type: number
            example: 31000.13
          otrasdietaseindemnizaciones_sueldo: 
            type: number
          retribucionmensual: 
            type: number
            example: 1000
          retribucionanual:
            type: number
            example: 0
          observaciones: 
            type: string
            example: "Sin dedicación"
      xml:
        name: politician
    PoliticianResponse:
      type: object
      properties:
          _error: 
            type: string
            example: "Invalid politician ID"
          _index: 
            type: string
            example: "politiciansIndex"
          _id:
            type: string
            example: "12345ABCDEF"
          _version: 
            type: number
            example: 1
          _seq_no: 
            type: number
            example: 0
          _primary_term: 
            type: number
            example: 2
          found: 
            type: boolean
          _score: 
            type: number
            example: 0.76
          _source:
            $ref: '#/components/schemas/Politician'
      xml:
        name: politicianResponse
    PoliticianPatchResponse:
      type: object
      properties:
          _error: 
            type: string
            example: "Invalid politician ID"
          _index: 
            type: string
            example: "politiciansIndex"
          _id:
            type: string
            example: "12345ABCDEF"
          _version: 
            type: number
            example: 1
          result: 
            type: string
            example: "updated"
          _shards: 
            type: object
          _seq_no: 
            type: number
            example: 4097
          _primary_term: 
            type: number
            example: 2
      xml:
        name: politicianPatchResponse
...
    ApiResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid politician ID"
      xml:
        name: '##default'